clean up transfer_check.py
